Create notification package in workers directory, move local notification classes
