+++ WordPress/src/main/java/org/wordpress/android/workers/notification/createsite/CreateSiteNotificationHandler.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.workers
+package org.wordpress.android.workers.notification.createsite
 
 import android.content.Context
 import android.content.Intent
@@ -7,6 +7,7 @@ import org.wordpress.android.fluxc.store.SiteStore
 import org.wordpress.android.push.NotificationType.CREATE_SITE
 import org.wordpress.android.ui.ActivityLauncher
 import org.wordpress.android.ui.notifications.SystemNotificationsTracker
+import org.wordpress.android.workers.notification.local.LocalNotificationHandler
 import javax.inject.Inject
 
 class CreateSiteNotificationHandler @Inject constructor(
