+++ WordPress/src/main/java/org/wordpress/android/workers/notification/local/LocalNotification.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.workers
+package org.wordpress.android.workers.notification.local
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
