+++ WordPress/src/main/java/org/wordpress/android/workers/notification/local/LocalNotificationHandlerFactory.kt
@@ -1,9 +1,10 @@
-package org.wordpress.android.workers
+package org.wordpress.android.workers.notification.local
 
 import android.content.Context
 import android.content.Intent
-import org.wordpress.android.workers.LocalNotification.Type
-import org.wordpress.android.workers.LocalNotification.Type.CREATE_SITE
+import org.wordpress.android.workers.notification.createsite.CreateSiteNotificationHandler
+import org.wordpress.android.workers.notification.local.LocalNotification.Type
+import org.wordpress.android.workers.notification.local.LocalNotification.Type.CREATE_SITE
 import javax.inject.Inject
 
 class LocalNotificationHandlerFactory @Inject constructor(
