+++ WordPress/src/main/java/org/wordpress/android/workers/notification/local/LocalNotificationWorker.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.workers
+package org.wordpress.android.workers.notification.local
 
 import android.app.PendingIntent
 import android.content.Context
@@ -12,7 +12,7 @@ import androidx.work.WorkerFactory
 import androidx.work.WorkerParameters
 import androidx.work.workDataOf
 import org.wordpress.android.R
-import org.wordpress.android.workers.LocalNotification.Type
+import org.wordpress.android.workers.notification.local.LocalNotification.Type
 
 class LocalNotificationWorker(
     val context: Context,
