+++ WordPress/src/test/java/org/wordpress/android/workers/CreateSiteNotificationHandlerTest.kt
@@ -12,6 +12,7 @@ import org.wordpress.android.fluxc.store.AccountStore
 import org.wordpress.android.fluxc.store.SiteStore
 import org.wordpress.android.push.NotificationType.CREATE_SITE
 import org.wordpress.android.ui.notifications.SystemNotificationsTracker
+import org.wordpress.android.workers.notification.createsite.CreateSiteNotificationHandler
 
 @RunWith(MockitoJUnitRunner::class)
 class CreateSiteNotificationHandlerTest {
